refactor(auth): rename package
