+++ cognito-idp/src/androidMain/kotlin/com/liftric/cognito/idp/jwt/Base64.kt
@@ -1,0 +1,15 @@
+package com.liftric.cognito.idp.jwt
+
+import android.util.Base64
+
+internal actual class Base64 {
+    actual companion object {
+        actual fun decode(input: String): String? {
+            return try {
+                String(Base64.decode(input, Base64.URL_SAFE), Charsets.UTF_8)
+            } catch (e: Exception) {
+                null
+            }
+        }
+    }
+}
