--- cognito-idp/src/androidTest/kotlin/com/liftric/cognito-idp/IdentityProviderClientTests.kt
@@ -1,12 +1,0 @@
-package com.liftric.auth
-
-import kotlinx.coroutines.runBlocking
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-
-@RunWith(RobolectricTestRunner::class)
-actual class IdentityProviderClientTests: AbstractIdentityProviderClientTests()
-
-actual fun runTest(block: suspend () -> Unit) = runBlocking {
-    block.invoke()
-}
