--- cognito-idp/src/commonMain/kotlin/com/liftric/cognito-idp/jwt/Base64.kt
@@ -1,7 +1,0 @@
-package com.liftric.auth.jwt
-
-internal expect class Base64 {
-    companion object {
-        fun decode(input: String): String?
-    }
-}
