+++ cognito-idp/src/iosTest/kotlin/com/liftric/cognito/idp/IdentityProviderClientTests.kt
@@ -1,0 +1,9 @@
+package com.liftric.cognito.idp
+
+import kotlinx.coroutines.runBlocking
+
+actual class IdentityProviderClientTests: AbstractIdentityProviderClientTests()
+
+actual fun runTest(block: suspend () -> Unit) = runBlocking {
+    block.invoke()
+}
