--- cognito-idp/src/jvmTest/kotlin/com/liftric/cognito-idp/IdentityProviderClientTests.kt
@@ -1,9 +1,0 @@
-package com.liftric.auth
-
-import kotlinx.coroutines.runBlocking
-
-actual class IdentityProviderClientTests: AbstractIdentityProviderClientTests()
-
-actual fun runTest(block: suspend () -> Unit) = runBlocking {
-    block.invoke()
-}
