+++ cognito-idp/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.liftric.auth"/>
+<manifest package="com.liftric.cognito.idp"/>
